Make regex optional by moving ansi stripping behind feature flag
