+++ Cargo.toml
@@ -12,13 +12,14 @@ documentation = "https://docs.rs/console"
 readme = "README.md"
 
 [features]
-default = ["unicode-width"]
+default = ["unicode-width", "ansi-parsing"]
+ansi-parsing = ["regex"]
 
 [dependencies]
 clicolors-control = "1.0.1"
 lazy_static = "1"
 libc = "0.2"
-regex = { version = "1.3.1", default-features = false, features = ["std"] }
+regex = { version = "1.3.1", optional = true, default-features = false, features = ["std"] }
 unicode-width = { version = "0.1", optional = true }
 
 [target.'cfg(unix)'.dependencies]
